기존 roi와 roishot을 roi [subcommand] 형식으로 합침
기존 명령어와 비교해서 다음처럼 변경되었다.

roi -> roi server
roishot -> roi shot
